+++ src/Sharpy.Core/Partial.Str/Str.Sized.cs
@@ -7,10 +7,10 @@ public readonly partial struct Str
     /// <summary>
     /// Returns the number of (Unicode) characters in the string.
     /// </summary>
-    public uint __Len__()
+    public int __Len__()
     {
         StringInfo stringInfo = new(_s);
 
-        return (uint)stringInfo.LengthInTextElements;
+        return stringInfo.LengthInTextElements;
     }
 }
